RAWデータ表示の準備。

+++ app/src/main/java/net/osdn/gokigen/blecontrol/lib/data/brainwave/IBrainwaveDataReceiver.java
@@ -2,5 +2,5 @@ package net.osdn.gokigen.blecontrol.lib.data.brainwave;
 
 public interface IBrainwaveDataReceiver
 {
-    void receivedRawData(byte value);
+    void receivedRawData(int value);
 }
